# Conflicts:
#	src/main/java/org/buhuiqiming/fuchuang/entity/jpa/InterviewEntity.java

+++ src/main/java/org/buhuiqiming/fuchuang/dto/TurnEvaluationResult.java
@@ -25,7 +25,7 @@ public class TurnEvaluationResult implements Serializable {
     @Data
     public static class DimensionScores implements Serializable {
         private float professional;
-        private float cognition;
+        private Integer cognition;
         private float expression;
     }
